+++ src/Parser.js
@@ -1,4 +1,8 @@
-const jsrp = require('@apidevtools/json-schema-ref-parser')
+const asyncapiParser = require('@asyncapi/parser')
+const openapiSchemaParser = require('@asyncapi/openapi-schema-parser')
+const {promisify} = require('util')
+const readFile = promisify(require('fs').readFile)
+const ValidationError = require('./ValidationError')
 
 class Parser {
   /**
@@ -7,23 +11,30 @@ class Parser {
    */
   async parse(source) {
     try {
-      const options = {
-        parse: {
-          yaml: {
-            allowEmpty: false
-          },
-          text: false,
-          binary: false
-        },
-        dereference: {
-          circular: 'ignore'
-        }
+      asyncapiParser.registerSchemaParser(openapiSchemaParser)
+      if (source.indexOf('https://') === 0 || source.indexOf('http://') === 0) {
+        return await asyncapiParser.parseFromUrl(source)
       }
-      const schema = await jsrp.dereference(source, options)
-      return schema
+      const file = await readFile(source, 'utf8')
+      return await asyncapiParser.parse(file)
     } catch (err) {
-      throw new Error(err)
+      throw new ValidationError(this._formatError(err), undefined, err.validationErrors)
     }
+  }
+
+  /**
+   * @param {{ title: any; message: any; detail: any; validationErrors: any[]; }} err
+   */
+  _formatError(err) {
+    const title = err.title || err.message
+    let details = 'Error Details: '
+    details += err.detail ? err.detail : ''
+    if (err.validationErrors && err.validationErrors.length) {
+      err.validationErrors.forEach(element => {
+        details += element.title ? element.title : ''
+      })
+    }
+    return `${title} ${details}`
   }
 }